Fix manganelo grabber downloading (undesired) multiple chapters
Instead of searching for numbers only, the change searches for "Chapter
N" in the title, ignoring anything that's not passing that expression.

Since manganelo-like sites all have the "Chapter N" somewhere in their
title (except for some rare cases like announcements, which I'd say we
can ignore them safely), this should be safe for 99.99% of the cases.

refs #17